requirements.txt
anthropic>=0.40.0
httpx>=0.26.0
resend>=2.0.0
rich>=13.0
click>=8.0
python-dotenv>=1.0.0
pyyaml>=6.0
